+++ admin/models/translations.php
@@ -283,12 +283,15 @@ class GetbibleModelTranslations extends ListModel
 		}
 
 		// Add the list ordering clause.
-		$orderCol = $this->state->get('list.ordering', 'a.id');
-		$orderDirn = $this->state->get('list.direction', 'desc');
+		$orderCol = $this->state->get('list.ordering', 'a.language');
+		$orderDirn = $this->state->get('list.direction', 'asc');
 		if ($orderCol != '')
 		{
 			$query->order($db->escape($orderCol . ' ' . $orderDirn));
 		}
+
+		// Add a permanent list ordering.
+		$query->order($db->escape('a.translation asc'));
 
 		return $query;
 	}