+++ admin/views/translations/view.html.php
@@ -47,8 +47,8 @@ class GetbibleViewTranslations extends HtmlView
 		// Load the active filters.
 		$this->activeFilters = $this->get('ActiveFilters');
 		// Add the list ordering clause.
-		$this->listOrder = $this->escape($this->state->get('list.ordering', 'a.id'));
-		$this->listDirn = $this->escape($this->state->get('list.direction', 'DESC'));
+		$this->listOrder = $this->escape($this->state->get('list.ordering', 'a.language'));
+		$this->listDirn = $this->escape($this->state->get('list.direction', 'asc'));
 		$this->saveOrder = $this->listOrder == 'a.ordering';
 		// set the return here value
 		$this->return_here = urlencode(base64_encode((string) JUri::getInstance()));